Fix issue #12 - improve recovery process when scripts fail
This change reverts back to dbdeploy v1 behaviour.  The start time of executing a script is no longer recorded, and dbdeploy v2 users will have to manually remove the start_dt column from their changelog tables.

However, because the insert to changelog is done *after* the script itself was executed and within the same transaction, if there is an error in the change script the state of the database is clear: no trance of the change script has been applied.

Note that most database do not apply DDL transactionally.  To cope with DDL failure, the recommended approach is to include only ONE DDL statement per change script.  Or pester your database vendor to apply DDL transactionally.